backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
anthropic==0.34.2
python-dotenv==1.0.1
pillow==10.4.0
aiohttp==3.10.5
python-multipart==0.0.9
numpy==1.26.4
svgpathtools==1.6.1
